extracts menu item components to serparte file
